+++ HammerModel/Model/Units/Axis.cs
@@ -4,9 +4,9 @@ namespace HammerModel.Model.Units
 {
     public class Axis
     {
-        public int X { get; set; }
-        public int Y { get; set; }
-        public int Z { get; set; }
+        public double X { get; set; }
+        public double Y { get; set; }
+        public double Z { get; set; }
         public double Translation { get; set; }
         public double Factor { get; set; }
         public override string ToString()
